requirements.txt
gradio==3.50.2
together==0.2.7
python-dotenv==1.0.0